Vulkan 20 of X 01/06/2020 | 20w01a3 | 0.7.17

+++ TRAP/src/Graphics/API/Vulkan/VulkanShaderReflection.h
@@ -5,7 +5,7 @@
 
 namespace TRAP::Graphics::API
 {
-	ShaderReflection::ShaderReflection VkCreateShaderReflection(const std::vector<uint8_t>& shaderCode, RendererAPI::ShaderStage shaderStage);
+	ShaderReflection::ShaderReflection VkCreateShaderReflection(const std::vector<uint32_t>& shaderCode, RendererAPI::ShaderStage shaderStage);
 }
 
 #endif /*_TRAP_VULKANSHADERREFLECTION_H_*/
